requirements.txt
# Core LangChain
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1

# Modèles & Embeddings
requests==2.31.0
sentence-transformers==2.7.0  # Alternative locale

# Bases vectorielles
faiss-cpu==1.8.0  # Pour CPU
# faiss-gpu==1.8.0  # Option GPU

# Chargement de documents
beautifulsoup4==4.12.3  # HTML
pypdf==4.2.0  # PDF
unstructured[all-docs]==0.17.2  # Pour tous les types de documents
unstructured[pdf]

# Utilitaires
tiktoken==0.6.0  # Comptage de tokens
python-dotenv==1.0.1  # Gestion des variables d'environnement
numpy==1.26.4  # Calculs vectoriels

# API (optionnel)
fastapi==0.110.0  # Pour endpoint REST
uvicorn==0.27.1  # Serveur ASGI

# Tests & Qualité
pytest==8.0.2
pytest-cov==4.1.0
mypy==1.8.0
python-json-logger==2.0.7